crud workprocesses ready for testing
ik heb de create en update van de werkprocessen af.  de controller daarvoor aangepast. ik heb de styling in public gezet voor een globale css bestand. ook een js bestand in public gezet. heb wat dingen naar nederlands verandert. heb de style.css en script.js in elke index.blad gereferenced. create.blade voor werkprocessen gemaakt. een filter op kerntaken toegevoegd aan de werkprocess index. Een timer van 4 seconden op de succes messages die je krijgt als je iets create, update of delete. en de web.php aangepast voor de create pagina

+++ database/migrations/2023_12_12_143056_create_workprocess_tables.php
@@ -14,7 +14,7 @@ return new class extends Migration
     {
         Schema::create('workprocesses', function (Blueprint $table) {
             $table->id();
-            $table->integer('crebo');
+            $table->bigInteger('crebo');
             $table->foreignidFor(Core_task::class);
             $table->string('workprocess_number', 10);
             $table->string('workprocess_title');
